+++ test/fi/hut/soberit/agilefant/util/ExactEstimateUtilsTest.java
@@ -8,43 +8,43 @@ public class ExactEstimateUtilsTest {
 
     @Test
     public void testValidHours() {
-        assertEquals(60, ExactEstimateUtils.convertFromString("1h")
+        assertEquals((long)60, ExactEstimateUtils.convertFromString("1h")
                 .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursWithWhitespace() {
-        assertEquals(60, ExactEstimateUtils.convertFromString("  1h   ")
+        assertEquals((long)60, ExactEstimateUtils.convertFromString("  1h   ")
                 .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursWithDecimalsAsComma() {
-        assertEquals(90, ExactEstimateUtils.convertFromString("1,5h")
+        assertEquals((long)90, ExactEstimateUtils.convertFromString("1,5h")
                 .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursWithDecimalsAsPeriod() {
-        assertEquals(90, ExactEstimateUtils.convertFromString("1.5h")
+        assertEquals((long)90, ExactEstimateUtils.convertFromString("1.5h")
                 .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidMinutes() {
-        assertEquals(15, ExactEstimateUtils.convertFromString("15min")
+        assertEquals((long)15, ExactEstimateUtils.convertFromString("15min")
                 .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursAndMinutes() {
-        assertEquals(105, ExactEstimateUtils.convertFromString("1.5h15min")
+        assertEquals((long)105, ExactEstimateUtils.convertFromString("1.5h15min")
                 .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursAndMinutesWithWhitespace() {
-        assertEquals(105, ExactEstimateUtils.convertFromString(
+        assertEquals((long)105, ExactEstimateUtils.convertFromString(
                 "   1.5h   15min        ").getMinorUnits().longValue());
     }
